Added profile view
