+++ client/src/App.js
@@ -3,10 +3,10 @@ import React from "react";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 import LogoBar from "./components/navbar/LogoBar";
 import HomePage from './views/HomePage';
-import Profile from './views/Profile';
 import ViewProductDetail from './components/viewProduct/ViewProductDetail'
 import CreateProduct from './views/CreateProductDescription'
 import MessageView from './views/MessageView'
+import Profile from './views/Profile'
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 function App() {
@@ -24,7 +24,7 @@ function App() {
           <CreateProduct />
         </Route>
         <Route path="/Profile">
-          <Profile />
+        <Profile />
         </Route>
         <Route path="/">
           <HomePage />
